Hackathon updates

+++ starter-artifacts/deploy/powershell/Publish-Site.ps1
@@ -11,7 +11,11 @@ Push-Location $(./Join-Path-Recursively.ps1 -pathParts "..,..,ui".Split(","))
 Write-Host "===========================================================" -ForegroundColor Yellow
 Write-Host " Building the website" -ForegroundColor Yellow
 Write-Host "===========================================================" -ForegroundColor Yellow
-Remove-Item -Path ./out -Recurse -Force
+if (Test-Path ./out)
+{
+   Remove-Item -Path ./out -Recurse -Force
+}
+Start-Sleep -Seconds 10
 npm ci
 npm run build
 
@@ -19,6 +23,12 @@ Write-Host "===========================================================" -Foregr
 Write-Host " Deploying the website" -ForegroundColor Yellow
 Write-Host "===========================================================" -ForegroundColor Yellow
 az storage azcopy blob upload -c `$web --account-name $storageAccount -s "./out/*" --recursive
+
+$webUri=(az storage account show --name $storageAccount --resource-group $resourceGroup --query "primaryEndpoints.web" -o tsv)
+
+Write-Host "===========================================================" -ForegroundColor Yellow
+Write-Host " Website Uri: $webUri" -ForegroundColor Yellow
+Write-Host "===========================================================" -ForegroundColor Yellow
 
 Pop-Location
 Pop-Location
